refactor: Entity move
