--- src/main/java/com/tenbit/beep/attendance/domain/Attendance.java
@@ -1,9 +1,0 @@
-package com.tenbit.beep.attendance.domain;
-
-public enum Attendance {
-    TRUE,
-    FALSE
-}
-
-
-
