+++ src/main/java/com/tenbit/beep/auth/domain/Attendance.java
@@ -1,0 +1,6 @@
+package com.tenbit.beep.auth.domain;
+
+public enum Attendance {
+    ATTEND,
+    ABSENT
+}
