+++ src/main/java/com/tenbit/beep/classroom/dto/StudentInfoResponse.java
@@ -1,6 +1,6 @@
 package com.tenbit.beep.classroom.dto;
 
-import com.tenbit.beep.attendance.domain.Attendance;
+import com.tenbit.beep.auth.domain.Attendance;
 import com.tenbit.beep.auth.domain.User;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
